# enhancement: Apply perf patch to YAML parser