--- a/cmd/cerbosctl/internal/client/client.go
+++ b/cmd/cerbosctl/internal/client/client.go
@@ -25,7 +25,7 @@ func GetAdminClient(globals *flagset.Globals) (*cerbos.GRPCAdminClient, error) {
 
 	opts := globals.ToClientOpts()
 
-	ac, err := cerbos.NewAdminClientWithCredentials(globals.Server, globals.Username, globals.Password, opts...)
+	ac, err := cerbos.NewAdminClientWithCredentials("passthrough:///"+globals.Server, globals.Username, globals.Password, opts...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create the admin client: %w", err)
 	}
@@ -36,7 +36,7 @@ func GetAdminClient(globals *flagset.Globals) (*cerbos.GRPCAdminClient, error) {
 func GetClient(globals *flagset.Globals) (*cerbos.GRPCClient, error) {
 	opts := globals.ToClientOpts()
 
-	c, err := cerbos.New(globals.Server, opts...)
+	c, err := cerbos.New("passthrough:///"+globals.Server, opts...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create the client: %w", err)
 	}